requirements.txt
streamlit 
tinydb 
bcrypt 
gspread 
oauth2client 
qrcode 
opencv-python-headless 
numpy 
pillow
google-generativeai
python-dotenv